change file names + add project status info
file names cuz it was driving me crazy wdym reportbot_bot pls 😭
