+++ src/menus/report_bot.ts
@@ -8,7 +8,7 @@ import {
 } from 'discord.js'
 
 export const menu = {
-  name: 'reportserver_server',
+  name: 'report_bot',
 }
 
 export const execute = async (
@@ -18,16 +18,16 @@ export const execute = async (
   if (!interaction.inCachedGuild()) return
 
   const modal = new ModalBuilder()
-    .setCustomId('modModal_reportserver') // modal custom id
-    .setTitle('Report a Top.gg server')
+    .setCustomId('modModal_reportbot') // modal custom id
+    .setTitle('Report a Top.gg bot')
 
   const entityID = new TextInputBuilder()
     .setCustomId('entityID')
-    .setLabel('Server link')
+    .setLabel('Bot link')
     .setStyle(TextInputStyle.Short)
     .setRequired(true)
     .setMaxLength(1000)
-    .setPlaceholder('E.g. https://top.gg/discord/servers/id')
+    .setPlaceholder('E.g. https://top.gg/bot/id')
 
   const reasonInput = new TextInputBuilder()
     .setCustomId('modReason')
@@ -35,7 +35,7 @@ export const execute = async (
     .setStyle(TextInputStyle.Paragraph)
     .setRequired(true)
     .setMaxLength(1000)
-    .setPlaceholder('E.g. spam in the description, invalid invite, etc.')
+    .setPlaceholder('E.g. spamming, invalid invite, breaking tos, etc.')
 
   const screenshotInput = new TextInputBuilder()
     .setCustomId('Screenshot')
